test(backend): improve JUnit tests
