+++ build.gradle.kts
@@ -57,10 +57,10 @@ dependencies {
 
   testImplementation(Libs.vertx_junit5)
   testImplementation(Libs.vertx_junit5_web_client)
-  testImplementation(kotlin("test"))
-  testImplementation(kotlin("test-junit"))
+  // testImplementation(kotlin("test"))
+  // testImplementation(kotlin("test-junit"))
   testRuntimeOnly(Libs.junit_platform_launcher) {
-    because("Needed to run tests IDEs that bundle an older version")
+    because("Needed to run tests in IDEs that bundle an older version of JUnit")
   }
   testImplementation(Libs.junit_jupiter_api)
   testImplementation(Libs.junit_jupiter_params)
