+++ src/test/kotlin/it/unibo/protelis/web/execution/simulated/SimulatedProtelisEngineTest.kt
@@ -2,7 +2,7 @@ package it.unibo.protelis.web.execution.simulated
 
 import it.unibo.protelis.web.execution.ProtelisObserver
 import it.unibo.protelis.web.execution.ProtelisUpdateMessage
-import kotlin.test.assertTrue
+import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
